# Conflicts:
#	Sources/Internal/Engine/Private/OsX/Window/WindowNativeBridgeOsX.mm


Former-commit-id: 248eeef481e1ff03648ad3abab432a8c2814e162

+++ Programs/ResourceEditor/Classes/Qt/DeviceInfo/MemoryTool/Widgets/MemProfWidget.cpp
@@ -119,9 +119,9 @@ void MemProfWidget::RealtimeToggled(bool checked)
     if (!profileSession->IsFileMode() && realtimeMode)
     {
         QCustomPlot* plot = ui->plot;
-        size_t ngraph = profileSession->AllocPoolCount();
+        int ngraph = static_cast<int>(profileSession->AllocPoolCount());
 
-        for (size_t i = 0; i < ngraph; ++i)
+        for (int i = 0; i < ngraph; ++i)
         {
             QCPGraph* graph = plot->graph(i);
             graph->rescaleAxes(i > 0);
@@ -209,9 +209,9 @@ void MemProfWidget::SnapshotList_OnDoubleClicked(const QModelIndex& index)
 void MemProfWidget::UpdatePlot(const MemoryStatItem& stat)
 {
     QCustomPlot* plot = ui->plot;
-    size_t ngraph = profileSession->AllocPoolCount();
+    int ngraph = static_cast<int>(profileSession->AllocPoolCount());
 
-    for (size_t i = 0; i < ngraph; ++i)
+    for (int i = 0; i < ngraph; ++i)
     {
         QCPGraph* graph = plot->graph(i);
 
@@ -233,11 +233,11 @@ void MemProfWidget::UpdatePlot(const MemoryStatItem& stat)
 
 void MemProfWidget::SetPlotData()
 {
-    const size_t ntrends = profileSession->AllocPoolCount();
+    const int ntrends = static_cast<int>(profileSession->AllocPoolCount());
     const size_t nstat = profileSession->StatCount();
     Vector<QCPDataMap*> trends;
     trends.reserve(ntrends);
-    for (size_t i = 0; i < ntrends; ++i)
+    for (int i = 0; i < ntrends; ++i)
     {
         trends.push_back(new QCPDataMap);
     }
@@ -246,7 +246,7 @@ void MemProfWidget::SetPlotData()
         const MemoryStatItem& item = profileSession->Stat(i);
         const Vector<AllocPoolStat>& vstat = item.PoolStat();
         double key = static_cast<double>(item.Timestamp() / 1000.0);
-        for (size_t j = 0; j < ntrends; ++j)
+        for (int j = 0; j < ntrends; ++j)
         {
             double value = static_cast<double>(vstat[j].allocByApp) / 1024.0 / 1024.0;
             trends[j]->insert(key, QCPData(key, value));
@@ -254,7 +254,7 @@ void MemProfWidget::SetPlotData()
     }
 
     QCustomPlot* plot = ui->plot;
-    for (size_t i = 0; i < ntrends; ++i)
+    for (int i = 0; i < ntrends; ++i)
     {
         QCPGraph* graph = plot->graph(i);
         graph->setData(trends[i], false);
